Move f-string parser into own module and clean up a bit
