+++ parser/src/lib.rs
@@ -3,6 +3,7 @@ extern crate log;
 
 pub mod ast;
 pub mod error;
+mod fstring;
 pub mod lexer;
 pub mod parser;
 #[cfg_attr(rustfmt, rustfmt_skip)]
