+++ parser/src/parser.rs
@@ -65,180 +65,12 @@ pub fn parse_expression(source: &str) -> Result<ast::Expression, ParseError> {
     do_lalr_parsing!(source, Expression, StartExpression)
 }
 
-// TODO: consolidate these with ParseError
-#[derive(Debug, PartialEq)]
-pub enum FStringError {
-    UnclosedLbrace,
-    UnopenedRbrace,
-    InvalidExpression,
-}
-
-impl From<FStringError>
-    for lalrpop_util::ParseError<lexer::Location, token::Tok, lexer::LexicalError>
-{
-    fn from(_err: FStringError) -> Self {
-        lalrpop_util::ParseError::User {
-            error: lexer::LexicalError::StringError,
-        }
-    }
-}
-
-enum ParseState {
-    Text {
-        content: String,
-    },
-    FormattedValue {
-        expression: String,
-        spec: Option<String>,
-        depth: usize,
-    },
-}
-
-pub fn parse_fstring(source: &str) -> Result<ast::StringGroup, FStringError> {
-    use self::ParseState::*;
-
-    let mut values = vec![];
-    let mut state = ParseState::Text {
-        content: String::new(),
-    };
-
-    let mut chars = source.chars().peekable();
-    while let Some(ch) = chars.next() {
-        state = match state {
-            Text { mut content } => match ch {
-                '{' => {
-                    if let Some('{') = chars.peek() {
-                        chars.next();
-                        content.push('{');
-                        Text { content }
-                    } else {
-                        if !content.is_empty() {
-                            values.push(ast::StringGroup::Constant { value: content });
-                        }
-
-                        FormattedValue {
-                            expression: String::new(),
-                            spec: None,
-                            depth: 0,
-                        }
-                    }
-                }
-                '}' => {
-                    if let Some('}') = chars.peek() {
-                        chars.next();
-                        content.push('}');
-                        Text { content }
-                    } else {
-                        return Err(FStringError::UnopenedRbrace);
-                    }
-                }
-                _ => {
-                    content.push(ch);
-                    Text { content }
-                }
-            },
-
-            FormattedValue {
-                mut expression,
-                mut spec,
-                depth,
-            } => match ch {
-                ':' if depth == 0 => FormattedValue {
-                    expression,
-                    spec: Some(String::new()),
-                    depth,
-                },
-                '{' => {
-                    if let Some('{') = chars.peek() {
-                        expression.push_str("{{");
-                        chars.next();
-                        FormattedValue {
-                            expression,
-                            spec,
-                            depth,
-                        }
-                    } else {
-                        expression.push('{');
-                        FormattedValue {
-                            expression,
-                            spec,
-                            depth: depth + 1,
-                        }
-                    }
-                }
-                '}' => {
-                    if let Some('}') = chars.peek() {
-                        expression.push_str("}}");
-                        chars.next();
-                        FormattedValue {
-                            expression,
-                            spec,
-                            depth,
-                        }
-                    } else if depth > 0 {
-                        expression.push('}');
-                        FormattedValue {
-                            expression,
-                            spec,
-                            depth: depth - 1,
-                        }
-                    } else {
-                        values.push(ast::StringGroup::FormattedValue {
-                            value: Box::new(match parse_expression(expression.trim()) {
-                                Ok(expr) => expr,
-                                Err(_) => return Err(FStringError::InvalidExpression),
-                            }),
-                            spec: spec.unwrap_or_default(),
-                        });
-                        Text {
-                            content: String::new(),
-                        }
-                    }
-                }
-                _ => {
-                    if let Some(spec) = spec.as_mut() {
-                        spec.push(ch)
-                    } else {
-                        expression.push(ch);
-                    }
-                    FormattedValue {
-                        expression,
-                        spec,
-                        depth,
-                    }
-                }
-            },
-        };
-    }
-
-    match state {
-        Text { content } => {
-            if !content.is_empty() {
-                values.push(ast::StringGroup::Constant { value: content })
-            }
-        }
-        FormattedValue { .. } => {
-            return Err(FStringError::UnclosedLbrace);
-        }
-    }
-
-    Ok(match values.len() {
-        0 => ast::StringGroup::Constant {
-            value: String::new(),
-        },
-        1 => values.into_iter().next().unwrap(),
-        _ => ast::StringGroup::Joined { values },
-    })
-}
-
 #[cfg(test)]
 mod tests {
     use super::ast;
     use super::parse_expression;
-    use super::parse_fstring;
     use super::parse_program;
     use super::parse_statement;
-    use super::FStringError;
     use num_bigint::BigInt;
 
     #[test]
@@ -628,57 +460,6 @@ mod tests {
                     }
                 ],
             }
-        );
-    }
-
-    fn mk_ident(name: &str) -> ast::Expression {
-        ast::Expression::Identifier {
-            name: name.to_owned(),
-        }
-    }
-
-    #[test]
-    fn test_parse_fstring() {
-        let source = String::from("{a}{ b }{{foo}}");
-        let parse_ast = parse_fstring(&source).unwrap();
-
-        assert_eq!(
-            parse_ast,
-            ast::StringGroup::Joined {
-                values: vec![
-                    ast::StringGroup::FormattedValue {
-                        value: Box::new(mk_ident("a")),
-                        spec: String::new(),
-                    },
-                    ast::StringGroup::FormattedValue {
-                        value: Box::new(mk_ident("b")),
-                        spec: String::new(),
-                    },
-                    ast::StringGroup::Constant {
-                        value: "{foo}".to_owned()
-                    }
-                ]
-            }
-        );
-    }
-
-    #[test]
-    fn test_parse_empty_fstring() {
-        assert_eq!(
-            parse_fstring(""),
-            Ok(ast::StringGroup::Constant {
-                value: String::new(),
-            }),
-        );
-    }
-
-    #[test]
-    fn test_parse_invalid_fstring() {
-        assert_eq!(parse_fstring("{"), Err(FStringError::UnclosedLbrace));
-        assert_eq!(parse_fstring("}"), Err(FStringError::UnopenedRbrace));
-        assert_eq!(
-            parse_fstring("{class}"),
-            Err(FStringError::InvalidExpression)
         );
     }
 }
